Adds mont tangle/att design, the implementation, and two runners.

See merge request applied-topology/tanglenomicon/core_lib!22

+++ source/computations/rational_data/comp_rational_data.h
@@ -70,7 +70,7 @@ extern "C"
      */
     typedef struct
     {
-        int (*storage_write)(char *key, char *index, char *value);
+        uint8_t (*storage_write)(char *key, char *index, char *value);
         const char *(*storage_read)(char *key, char *index);
         note_tv_t *tv_n;
         char *tv_str_buff;
